Extracted I2C code into separate files
