+++ 32blit-stm32/Inc/32blit_i2c.h
@@ -4,3 +4,6 @@
 //
 
 #pragma once
+
+extern void blit_i2c_tick();
+extern void blit_init_accelerometer();
